+++ benchmarks/distributed/rpc/parameter_server/configurations/data_configurations.json
@@ -3,8 +3,8 @@
         "data_class": "DummyData",
         "configurations": {
             "max_val": 1024,
-            "input_samples": 1024,
-            "input_dim": 1024,
+            "sample_count": 1024,
+            "sample_length": 1024,
             "sparsity_percentage": 20
         }
     },
