--- benchmarks/distributed/rpc/parameter_server/criterion_functions/__init__.py
@@ -1,5 +1,0 @@
-from .cel import cel
-
-create_criterion_map = {
-    "cel": cel
-}
