+++ benchmarks/distributed/rpc/parameter_server/data/DummyData.py
@@ -11,22 +11,22 @@ class DummyData(Dataset):
     def __init__(
         self,
         max_val: int,
-        input_samples: int,
-        input_dim: int,
+        sample_count: int,
+        sample_length: int,
         sparsity_percentage: int
     ):
         r"""
         A data class that generates random data.
         Args:
             max_val (int): the maximum value for an element
-            input_samples (int): count of training samples
-            input_dim (int): number of elements in a sample
+            sample_count (int): count of training samples
+            sample_length (int): number of elements in a sample
             sparsity_percentage (int): the percentage of
                 embeddings used by the input data in each iteration
         """
         self.max_val = max_val
-        self.input_samples = input_samples
-        self.input_dim = input_dim
+        self.input_samples = sample_count
+        self.input_dim = sample_length
         self.sparsity_percentage = sparsity_percentage
 
         def generate_input():
@@ -45,9 +45,9 @@ class DummyData(Dataset):
             return torch.from_numpy(np.array(data))
 
         self.input = generate_input()
-        self.target = torch.randint(0, max_val, [input_samples])
+        self.target = torch.randint(0, max_val, [sample_count])
         self.start = 0
-        self.end = input_samples
+        self.end = sample_count
 
     def __len__(self):
         return len(self.input)
