--- benchmarks/distributed/rpc/parameter_server/ddp_model_functions/__init__.py
@@ -1,5 +1,0 @@
-from .basic_ddp_model import basic_ddp_model
-
-create_ddp_model_map = {
-    "basic_ddp_model": basic_ddp_model
-}
