--- benchmarks/distributed/rpc/parameter_server/hookstates/__init__.py
@@ -1,5 +1,0 @@
-from .BasicHookState import BasicHookState
-
-hook_state_map = {
-    "BasicHookState": BasicHookState
-}
