--- benchmarks/distributed/rpc/parameter_server/iteration_step_functions/__init__.py
@@ -1,5 +1,0 @@
-from .basic_iteration_step import basic_iteration_step
-
-iteration_step_map = {
-    "basic_iteration_step": basic_iteration_step
-}
