--- benchmarks/distributed/rpc/parameter_server/optimizer_functions/__init__.py
@@ -1,5 +1,0 @@
-from .sgd_optimizer import sgd_optimizer
-
-create_optimizer_map = {
-    "sgd_optimizer": sgd_optimizer
-}
