--- benchmarks/distributed/rpc/parameter_server/optimizer_functions/sgd_optimizer.py
@@ -1,11 +1,0 @@
-import torch
-
-
-def sgd_optimizer(parameters, lr):
-    r"""
-    A function that creates a SGD optimizer for training.
-    Args:
-        parameters (iterable): iterable of parameters to optimize
-        lr (float): learning rate
-    """
-    return torch.optim.SGD(parameters, lr)
