--- benchmarks/distributed/rpc/parameter_server/preprocess_data_functions/__init__.py
@@ -1,5 +1,0 @@
-from .preprocess_dummy_data import preprocess_dummy_data
-
-preprocess_data_map = {
-    "preprocess_dummy_data": preprocess_dummy_data
-}
