+++ benchmarks/distributed/rpc/parameter_server/trainer/ddp_models.py
@@ -1,7 +1,7 @@
 from torch.nn.parallel import DistributedDataParallel as DDP
 
 
-def basic_ddp_model(self, rank, model, process_group, HookState, hook):
+def basic_ddp_model(self, rank, model, process_group, hook_state, hook):
     r"""
     A function that creates a ddp_model and hook_state objects.
     The ddp model is is initialized with a single device id and
@@ -18,6 +18,6 @@ def basic_ddp_model(self, rank, model, process_group, HookState, hook):
     ddp_model = DDP(
         model, device_ids=[rank], process_group=process_group
     )
-    hook_state = HookState(self, process_group)
+    hook_state = hook_state(self, process_group)
     ddp_model.register_comm_hook(hook_state, hook)
     return ddp_model, hook_state
