--- benchmarks/distributed/rpc/parameter_server/trainers/__init__.py
@@ -1,5 +1,0 @@
-from .DdpTrainer import DdpTrainer
-
-trainer_map = {
-    "DdpTrainer": DdpTrainer
-}
